% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frbca.R
\name{apply_eal_weighting}
\alias{apply_eal_weighting}
\title{Apply weighting to EAL data by dividing loss by specified column}
\usage{
apply_eal_weighting(data, w = NULL)
}
\arguments{
\item{data}{Data frame from postprocess_eal with loss column}

\item{w}{Column name to divide loss by (NULL for no weighting)}
}
\value{
Data frame with weighted loss and attribute for y-label
}
\description{
Apply weighting to EAL data by dividing loss by specified column
}
