% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frbca.R
\name{frbca}
\alias{frbca}
\title{Conduct FR-BCA with sensitivity analysis}
\usage{
frbca(eal, cost, params)
}
\arguments{
\item{eal}{Table of Expected Annualized Losses (EAL)}

\item{cost}{Table of Structural, Nonstructural, and Total Construction Costs}

\item{params}{List of analysis parameters}
}
\value{
Table including EAL, Costs, Benefits, FR-BCA Outputs (BCR and NPV)
}
\description{
Wrapper that conducts BCA for each set of models in list
}
