% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frbca.R
\name{plot_bcr}
\alias{plot_bcr}
\title{Plot FR-BCA Outputs}
\usage{
plot_bcr(output, n_floors = 4, system = "RCMF")
}
\arguments{
\item{output}{Output from \code{frbca()}}

\item{n_floors}{Number of stories (for figure title)}

\item{system}{Name of structural system being plotted (default: "RCMF")}
}
\value{
Updated model table including PV(Cost)
}
\description{
Sensitivity analysis plot for FR-BCA outputs, for fixed story height and structural system
}
