% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frbca.R
\name{plot_bcr_sensitivity}
\alias{plot_bcr_sensitivity}
\title{Plot FR-BCA Outputs with Sensitivity Analysis}
\usage{
plot_bcr_sensitivity(
  output,
  systems = "RCMF",
  designs = "nonstructural",
  stories = 4
)
}
\arguments{
\item{output}{Output from \code{frbca()}}

\item{n_floors}{Number of stories (for figure title)}

\item{system}{Name of structural system being plotted (default: "RCMF")}
}
\value{
Updated model table including PV(Cost)
}
\description{
Sensitivity analysis plot for FR-BCA outputs, for fixed story height and structural system
}
