% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frbca.R
\name{plot_cost_delta}
\alias{plot_cost_delta}
\title{Plot Cost deltas for recovery-based design interventions}
\usage{
plot_cost_delta(
  output,
  systems = "RCMF",
  designs = "nonstructural",
  stories = 4
)
}
\description{
Plot Cost deltas for recovery-based design interventions
}
