% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frbca.R
\name{plot_eal}
\alias{plot_eal}
\title{Plot Total EALs across multiple systems and stories}
\usage{
plot_eal(
  output,
  systems = "RCMF",
  designs = "nonstructural",
  stories = 4,
  w = FALSE
)
}
\description{
Plot Total EALs across multiple systems and stories
}
