% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frbca.R
\name{plot_eal_by_loss}
\alias{plot_eal_by_loss}
\title{Plot EALs disaggregated by loss category}
\usage{
plot_eal_by_loss(
  output,
  systems = "RCMF",
  designs = "nonstructural",
  stories = 4
)
}
\description{
Plot EALs disaggregated by loss category
}
