% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frbca.R
\name{plot_eal_weighted}
\alias{plot_eal_weighted}
\title{Plot Weighted EALs by dividing loss by specified column}
\usage{
plot_eal_weighted(
  output,
  systems = "RCMF",
  designs = "nonstructural",
  stories = 4,
  w = NULL
)
}
\description{
Plot Weighted EALs by dividing loss by specified column
}
