% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frbca.R
\name{plot_frbca}
\alias{plot_frbca}
\title{Plot FR-BCA Outputs}
\usage{
plot_frbca(output, n_floors = 4, system = "RCMF")
}
\arguments{
\item{model}{A model}

\item{params}{A list of parameters}
}
\value{
Updated model table including PV(Cost)
}
\description{
Sensitivity analysis plot for FR-BCA outputs
}
