% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frbca.R
\name{pv_cost}
\alias{pv_cost}
\title{Present Value Cost}
\usage{
pv_cost(model, params)
}
\arguments{
\item{model}{A model}

\item{params}{A list of parameters}
}
\value{
Updated model table including PV(Cost)
}
\description{
Compute the present value (PV) structural, nonstructural, and total construction cost
}
